.github/auto-translation/pyproject.toml
[project]
name = "jhipster-auto-translation"
version = "0.1.0"
description = "JHipster日本語ドキュメント自動翻訳システム"
authors = [
    {name = "JHipster Team"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "pytest (>=8.4.0,<9.0.0)",
    "language-tool-python (>=2.9.4,<3.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "gitpython (>=3.1.44,<4.0.0)",
    "requests (>=2.32.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
